[bugfix]modified exporting sequence data not right when the max/min vaule is not specified
Match-id-6d0cf9d144e8f7d8844525302b7d7e77e2f1a5f8

+++ code/datastudio/src/com.huawei.mppdbide.bl/src/com/huawei/mppdbide/bl/export/ExportObjectDataManager.java
@@ -120,8 +120,8 @@ public class ExportObjectDataManager {
             this.stmt = getExportConn().createStatement();
             rs = stmt.executeQuery(seqMaxQry);
             while (rs.next()) {
-                maxValue = rs.getInt("max_value");
-                minValue = rs.getInt("min_value");
+                maxValue = rs.getLong("max_value");
+                minValue = rs.getLong("min_value");
                 called = rs.getBoolean("is_called");
             }
             nextVal = getSeqNextValue(seqObject);
